Fix tests

+++ tests/data/schema_tests.py
@@ -20,7 +20,7 @@ import tensorfx as tfx
 class TestCases(unittest.TestCase):
 
   def test_create_single_field_schema(self):
-    f = tfx.data.SchemaField.integer('n')
+    f = tfx.data.SchemaField.numeric('n')
     schema = tfx.data.Schema.create(f)
 
     self.assertEqual(len(schema), 1)
@@ -28,7 +28,7 @@ class TestCases(unittest.TestCase):
     self.assertEqual(schema[0], f)
 
   def test_create_multi_field_schema(self):
-    f1 = tfx.data.SchemaField.integer('n')
+    f1 = tfx.data.SchemaField.numeric('n')
     f2 = tfx.data.SchemaField.text('t')
     schema = tfx.data.Schema.create(f1, f2)
 
@@ -40,7 +40,7 @@ class TestCases(unittest.TestCase):
     spec = """
     fields:
     - name: f1
-      type: integer
+      type: numeric
     - name: f2
       type: text
     - name: f3
@@ -50,6 +50,6 @@ class TestCases(unittest.TestCase):
 
     self.assertEqual(len(schema), 3)
     self.assertEqual(schema[0].name, 'f1')
-    self.assertEqual(schema['f1'].type, tfx.data.SchemaFieldType.integer)
+    self.assertEqual(schema['f1'].type, tfx.data.SchemaFieldType.numeric)
     self.assertEqual(schema.fields, ['f1', 'f2', 'f3'])
